refactor(fakeCuratedProjects): Move fakeCuratedProjects to supabase mocks
